+++ HOMELY.py
@@ -3,32 +3,33 @@
 import sys
 import os
 
-from homely.files import mkdir, symlink
-from homely.system import haveexecutable, execute
-
+import homely.files as files
+import homely.system as system
 
 HERE = os.path.dirname(os.path.realpath(__file__))
 
 
 def main():
-    symlink("bashrc", ".bashrc")
-    symlink("bash_profile", ".bash_profile")
-    symlink("profile", ".profile")  # TODO maybe add lines instead?
-    symlink("vimrc", ".vimrc")
-    # TODO check if taskwarrior is installed and only symlink if so
-    symlink("taskrc", ".taskrc")
+    files.symlink("bashrc", ".bashrc")
+    files.symlink("bash_profile", ".bash_profile")
+    files.symlink("profile", ".profile")  # TODO maybe add lines instead?
+    if system.haveexecutable("task"):
+        files.symlink("taskrc", ".taskrc")
 
     # git common config:
     # TODO error if git not present
     # TODO into separate function
     gitpath = os.path.join(HERE, "common.gitconfig")
-    execute(["git", "config", "--global", "include.path", gitpath])
+    system.execute(["git", "config", "--global", "include.path", gitpath])
     # NOTE this might overwrite if there's another include.path already present
     # TODO revert/cleanup somehow?
     # TODO check that path specification using ~/ works
 
-    mkdir(".git-hooks")
-    symlink("post-checkout", ".git-hooks/post-checkout")
+    files.mkdir(".git-hooks")
+    files.symlink("post-checkout", ".git-hooks/post-checkout")
+
+    files.symlink("vimrc", ".vimrc")
+    # install vim-plug
 
     return 0