+++ CHANGELOG.rst
@@ -10,6 +10,11 @@ and this project adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.
 Unreleased_
 ===========
 
+Changed
+-------
+
+* Only install .taskrc if taskwarrior is present
+
 0.1.0 - 2018-08-05
 ==================